+++ Assets/Core/Game/Domains/GamePlay/Simulation/Scripts/MatchModel/MatchDataService.cs
@@ -73,7 +73,7 @@ namespace Core.Game.Domains.GamePlay.Simulation.Scripts.MatchModel
             wallState.Points = wallPoints;
         }
         
-        public TalentCardS2C AddTalentCard(ushort talentCardId, Vector2 position, TalentType talentType, int Health)
+        public TalentCardS2C AddTalentCard(ushort talentCardId, Vector2 position, TalentType talentType, ushort Health)
         {
             ref var newCard = ref _simulationState.TalentCards.AddAndGet();
             newCard.Id = talentCardId;
